% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.flows}
\alias{df.flows}
\title{Mean monthly flows for USGS gauge 11169500
Naturalized flows from nhdPlus}
\format{
A data frame with 12 monthly values of flow
}
\usage{
df.flows
}
\description{
See https://edm44.github.io/hydr-watres-book/the-hydrologic-cycle-and-precipitation.html
usethis::use_data(df.flows)
}
\keyword{datasets}
