% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.p_prism}
\alias{df.p_prism}
\title{Mean monthly PRISM precipitation in mm for the basin draining to USGS gauge 11169500
downloaded using R package prism}
\format{
A data frame with 12 monthly values of precipitation
}
\usage{
df.p_prism
}
\description{
See https://edm44.github.io/hydr-watres-book/the-hydrologic-cycle-and-precipitation.html
usethis::use_data(df.p_prism)
}
\keyword{datasets}
