% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_functions.R
\name{DO_functions}
\alias{DO_functions}
\alias{Mixture}
\alias{Kadj_deox}
\alias{Kadj_reox}
\alias{O2sat}
\alias{DOdeficit}
\title{Functions to calculate a Streeter-Phelps dissolved oxygen sag curve}
\usage{
Mixture(Q1 = NULL, T1 = NULL, Q2 = NULL, T2 = NULL)

Kadj_deox(K1 = NULL, T = NULL)

Kadj_reox(K2 = NULL, T = NULL)

O2sat(T = NULL, P = 1)

DOdeficit(t = NULL, K1 = NULL, K2 = NULL, L0 = NULL, D0 = NULL)
}
\arguments{
\item{Q1}{first input flow rate to Mixture function [\eqn{m^3 s^{-1}}{m^3/s}}

\item{T1}{the first conservative constituent to the Mixture function, usually
a temperature or concentration}

\item{Q2}{the second input flow rate to Mixture function [\eqn{m^3 s^{-1}}{m^3/s}}

\item{T2}{the second conservative constituent to the Mixture function, usually
a temperature or concentration}

\item{K1}{deoxygenation coefficient (at 20\eqn{^{\circ}C}, base e) [\eqn{d^{-1}}{1/d}]}

\item{T}{the water temperature [\eqn{^{\circ}C}{C}]}

\item{K2}{reaeration coefficient (at 20\eqn{^{\circ}C}, base e) [\eqn{d^{-1}}{1/d}]}

\item{P}{atmospheric pressure (optional, default = 1) [atm]}

\item{t}{time [d]}

\item{L0}{the ultimate BOD (at t=infinity) [\eqn{mg L^{-1}}{mg/L}]}

\item{D0}{the initial DO deficit [\eqn{mg L^{-1}}{mg/L}]}
}
\value{
From the Mixture function: The concentration or temperature of the mixture

From the Kadj_deox and Kadj_reox functions: The adjusted K value [\eqn{d^{-1}}{1/d}]

From O2sat function, the saturation dissolved oxygen [\eqn{mg L^{-1}}{mg/L}]

from DOdeficit function, the dissolved oxygen at the given time(s) [\eqn{mg L^{-1}}{mg/L}]
}
\description{
These functions are used in a classroom exercise for dissolved oxygen
deficit calculations. Functions are `Mixture()` to calculate the resulting temperature
or concentration from two streams; `O2sat()` to calculate the saturation concentration
of oxygen; `DOdeficit()` to find the dissolved oxygen deficit at a given time; and
`Kadj_deox()` and `Kadj_reox()` to adjust deoxygenation and reaeration coefficients for
actual temperature prior to calculating dissolved oxygen deficit.
}
\author{
Ed Maurer
}
