% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{four_trees_swe}
\alias{four_trees_swe}
\title{Steps to produce four_trees_swe data frame
Snow Water Equivalent data accessed through
https://www.nrcs.usda.gov/wps/portal/wcc/}
\format{
A data frame with April 1 snow water equivalent values for 1980-2023
}
\usage{
four_trees_swe
}
\description{
Navigated to the snow data for
Four Trees (FOR) California  COOPERATOR SNOW SENSORS Site - 5150 ft#
Manually extracted April 1 data and renamed columns.
four_trees_swe <- readr::read_csv("./data-raw/four_trees_swe.csv")
usethis::use_data(four_trees_swe)
}
\keyword{datasets}
