% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrainSize.R
\name{GrainSize}
\alias{GrainSize}
\alias{grainSize.plot}
\alias{percentComponents}
\alias{percentSSC}
\alias{Dsize}
\alias{pctfromSize}
\alias{grainSize.coefs}
\title{Functions adapted from geotech package related to grainsize distributions}
\usage{
grainSize.plot(sieve = NULL, size = NULL, percent, metric = TRUE)

percentComponents(sieve = NULL, size = NULL, percent, metric = TRUE)

percentSSC(sieve = NULL, size = NULL, percent, metric = TRUE)

Dsize(N, sieve = NULL, size = NULL, percent, metric = TRUE)

pctfromSize(DN, sieve = NULL, size = NULL, percent, metric = TRUE)

grainSize.coefs(
  percent,
  sieve = NULL,
  size = NULL,
  D10 = NULL,
  D30 = NULL,
  D60 = NULL
)
}
\arguments{
\item{sieve}{an input vector of sieve numbers. Used if particle diameters are not provided}

\item{size}{an input numeric vector of particle sizes [mm or in]}

\item{percent}{the percent finer corresponding to each sieve or size [percent]}

\item{metric}{boolean for units of sizes: TRUE = mm, FALSE = in (Default = TRUE)}

\item{N}{a single percent finer used in the Dsize function [percent]}

\item{DN}{a single diameter used in the pctfromSize function [mm or in]}

\item{D10}{the diameter with 10 percent finer, optional argument to grainSize.coefs [mm or in]}

\item{D30}{the diameter with 30 percent finer, optional argument to grainSize.coefs [mm or in]}

\item{D60}{the diameter with 60 percent finer, optional argument to grainSize.coefs [mm or in]}
}
\value{
From the grainSize.coefs function: a list of the uniformity coefficient (Cu), curvature coefficient (Cc) and the effective size (D10) of the soil sample

From the grainSize.plot function: a grainsize distribution plot

From the percentComponents function: The percents of gravel, sand, and fines

From the Dsize function: The size associated with the given percent finer

From the pctfromSize function: The percent finer associated with the given diameter
}
\description{
These functions are portions of the `geotech` package used to
analyze and plot grainsize distributions for soils. Primary functions are
`grainSize.coefs()` to calculate distribution characteristics; `Dsize()` to
return the size associated with a percent finer; `pctfromSize()` to return
the percent finer for a given diameter; `percentComponents` to
calculate the percents of gravel, sand, and fines; `percentSSC()` to return
percents sand, silt, and clay (excluding gravel); grainSize.plot()` to create
a typical grainsize distribution plot. Most are adapted from the `geotech` package
code originally written by Kyle Elmy and Jim Kaklamanos, 2015.
}
