% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflows_20years}
\alias{inflows_20years}
\title{inflows_20years produced using the following
synth_mean <- c(NA)*12
inflows <- c(95,112,170,250,265,62,35,18,55,88,85,90)
sdi <- exp(.7*log(inflows))
inflows_synth <- rnorm(240,inflows,sdi)
inflows_synth <- ts(inflows_synth,start = c(2000, 1), frequency = 12)
synth_mean <- aggregate(c(inflows_synth), list(month = cycle(inflows_synth)), mean)
scale_ratios <- inflows/synth_mean$x
scaled_synth_flows <- inflows_synth*scale_ratios
inflows_20years <- ts(scaled_synth_flows,start = c(2000, 1), frequency = 12)}
\format{
A time series of 20 years of synthetic monthly streamflow in 10^6 m^3
}
\usage{
inflows_20years
}
\description{
inflows are from Wurbs and James, Water Resources Engineering, Example 11.7
}
\keyword{datasets}
