% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precip_double_mass}
\alias{precip_double_mass}
\title{USGS Double Mass curve data}
\format{
A data frame with year and two columns of precipitation
}
\usage{
precip_double_mass
}
\description{
Precipitation data from Table 1 from Double-Mass Curves,
USGS Geological Survey Water Supply Paper 1541-B, Washington, DC, 1960
}
\keyword{datasets}
