% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Qpeak_download}
\alias{Qpeak_download}
\title{USGS Peak flow data}
\format{
A data frame with annual peak flow values for station 11169000
}
\usage{
Qpeak_download
}
\description{
USGS peak flow data downloaded using this command
dataRetrieval::readNWISpeak(siteNumbers="11169000")
}
\keyword{datasets}
