% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scc_county}
\alias{scc_county}
\title{SCC County}
\format{
An sf object with 1 rows and many columns
}
\usage{
scc_county
}
\description{
A county boundary for Santa Clara County
}
\keyword{datasets}
