% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{snow}
\alias{snow}
\alias{snow.sim}
\title{Simple degree day factor snow model}
\usage{
snow.sim(
  DATA,
  Tmax,
  Tmin,
  kd,
  kdmonth = NULL,
  kf = 1,
  rcap = 0.025,
  Tmelt = 0,
  deltaT = 0,
  cr = 1,
  cs = 1,
  LSWE_0 = 0,
  ISWE_0 = 0
)
}
\arguments{
\item{DATA}{a \code{\link{ts}}-like object with named columns: \describe{
\item{list("P")}{ time series of areal rainfall depths, usually in mm. May be "precipitation". }
\item{list("T")}{ time series of temperature. May be "temperature_mean". } }}

\item{Tmax}{temperature threshold for rain, all precipitation is liquid above this
threshold.}

\item{Tmin}{temperature threshold for rain, all precipitation is snow below this
threshold.}

\item{kd}{degree day factor for snowmelt. Can be a numeric value or vector of length 2.}

\item{kdmonth}{if kd is a vector of 2 values, the month at which kd[2] applies. Ignored otherwise.}

\item{kf}{degree day factor for freezing (default = 1).}

\item{rcap}{retention parameter for liquid water capacity of snowpack (default = 0.025).}

\item{Tmelt}{temperature threshold for snowmelt and freezing in the
snowpack.}

\item{deltaT}{fixed temperature to add to all T values (default = 0)}

\item{cr}{correction factor for rainfall (default = 1).}

\item{cs}{correction factor for snowfall. (default = 1)}

\item{LSWE_0, ISWE_0}{initial values of state variables (defaults = 0).}
}
\value{
Returns a data frame containing:
\itemize{
  \item date
  \item swe_simulated - the simulated snow water equivalent
  \item snowmeltdischarge - the water discharged from the snow pack
  \item rain - the precipitation falling as rain
  \item snow - the precipitation falling as snow
}
}
\description{
Simple degree day factor snow model, modified by E. Maurer to account for
snowpack cold content, and to recognize data format from snotelr package.
Also added ability to enter two values for early and late season melt indices.
}
\details{
SWE snow water equivalent
ISWE water equivalent of ice in the snowpack
LSWE liquid water retained in the snowpack
}
\references{
Kokkonen T., Jakeman A.J, Koivusalo.H, Norton.J.: COMPUTATIONAL
METHODS FOR WATER RESOURCE ASSESSMENTS: AN EXERCISE KIT Educational Series
on Modelling and Software iEMSs International Modelling and Software Society
Available through www.iemss.org
}
\author{
Coded in R by Jarkko Koskela @tkk.fi 2010-02-26.

Converted to C by Felix Andrews \email{felix@nfrac.org}.
}
