% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow_data}
\alias{snow_data}
\title{Snow Water Equivalent from SNOTEL}
\format{
A data frame with daily values of snow water equivalent
}
\usage{
snow_data
}
\description{
SNOTEL data downloaded using this command
snotelr::snotel_download(site_id = 1050, internal = TRUE)
}
\keyword{datasets}
