% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow_mc_out}
\alias{snow_mc_out}
\title{Snow index model output from monte carlo simulation
snow_data used as observed swe and input precip and
temp to snow model}
\format{
A vector of RMSE for each of 10,000 simulation runs
}
\usage{
snow_mc_out
}
\description{
See https://edm44.github.io/hydr-watres-book/fate-of-precipitation.html
save(out, file = "snow_mc_out.rda")
}
\keyword{datasets}
