% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theis.R
\name{theis}
\alias{theis}
\alias{cooper.jacob}
\alias{theis_params}
\title{Solve for the groundwater parameters transmissivity and storativity based on
the Theis solution. Includes the simplifcation of Cooper Jacob.}
\usage{
cooper.jacob(Q, r, t, s)

theis_params(Q, r, t, s)
}
\arguments{
\item{Q}{Pumping rate [L^3/T]}

\item{r}{radial distance from pumping well to observation point [L]}

\item{t}{a vector of times [T]}

\item{s}{a numeric vector of drawdowns at each time t recorded at distance r [L]+}
}
\value{
a list of aquifer parameters: transmissivity, Tr [L^2/T] and storativity, St []
}
\description{
All code was adapted from the rhytool code produced by
Philippe Renard, François Bertone (https://github.com/FrancoisBertone/rhytool).
}
\details{
The original code uses a MIT license, which persists in this portion of the
code: Permission is hereby granted, free of charge, to any person obtaining a
copy of this software and associated documentation files (the "Software"), to
deal in the Software without restriction, including without limitation the
rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
sell copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:
The above copyright notice and this permission notice shall be included
in all copies or substantial portions of the Software.
}
\references{
Renard, Philippe (2017). Hytool: an open source matlab toolbox for the interpretation of hydraulic tests using analytical solutions. Journal of Open Source Software, 2(19), 441, \href{http://joss.theoj.org/papers/10.21105/joss.00441}{doi:10.21105/joss.00441}
}
